Refactor a bunch of code
